Fix coordinate-channel dtype in v100 compressor/flow processor
add_coord_channels built positional coords in float32 regardless of
the model's actual dtype; torch.cat then silently promoted the whole
tensor back to float32, breaking bf16/fp16 inference (e.g. ComfyUI's
new model-loader dtype option). Match dtype=x.dtype in both.

+++ src/fluxflow/models/v100/flow.py
@@ -336,8 +336,8 @@ class FluxFlowProcessor_v100(nn.Module):
         """Add normalized coordinate channels to feature map."""
         B, _, H, W = x.shape
         yy, xx = torch.meshgrid(
-            torch.linspace(-1, 1, H, device=x.device),
-            torch.linspace(-1, 1, W, device=x.device),
+            torch.linspace(-1, 1, H, device=x.device, dtype=x.dtype),
+            torch.linspace(-1, 1, W, device=x.device, dtype=x.dtype),
             indexing="ij",
         )
         coords = torch.stack([xx, yy], dim=0).unsqueeze(0).expand(B, -1, -1, -1)
